requirements.txt
aspose-words==22.12.0
cairocffi==1.4.0
CairoSVG==2.5.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
cssselect2==0.7.0
defusedxml==0.7.1
numpy==1.24.1
Pillow==9.4.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
tinycss2==1.2.1
tqdm==4.64.1
webencodings==0.5.1
